+++ app/src/main/java/com/unipet7/mcommerce/activities/SupportContact.java
@@ -4,29 +4,14 @@ import androidx.annotation.Nullable;
 import androidx.appcompat.app.ActionBar;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
-import androidx.fragment.app.Fragment;
 import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentPagerAdapter;
-import androidx.viewpager.widget.ViewPager;
 import androidx.viewpager2.widget.ViewPager2;
 
-import android.graphics.Color;
 import android.os.Bundle;
-import android.text.Spannable;
-import android.text.SpannableString;
-import android.text.style.TextAppearanceSpan;
-import android.view.Gravity;
-import android.widget.LinearLayout;
-import android.widget.TabHost;
-import android.widget.TextView;
 
 import com.google.android.material.tabs.TabLayout;
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.SupportViewPageAdapter;
-import com.unipet7.mcommerce.databinding.ActivityRateOrderBinding;
-import com.unipet7.mcommerce.databinding.ActivitySupportContactBinding;
-import com.unipet7.mcommerce.fragments.onboarding.ContactList;
-import com.unipet7.mcommerce.fragments.onboarding.Question;
 
 public class SupportContact extends AppCompatActivity {
     ActivitySupportContactBinding binding;
