+++ app/src/main/java/com/unipet7/mcommerce/fragments/ContactList.java
@@ -1,4 +1,4 @@
-package com.unipet7.mcommerce.fragments.onboarding;
+package com.unipet7.mcommerce.fragments;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -10,9 +10,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.databinding.FragmentContactListBinding;
-import com.unipet7.mcommerce.databinding.FragmentQuestionBinding;
 
 /**
  * A simple {@link Fragment} subclass.
