+++ app/src/main/java/com/unipet7/mcommerce/fragments/HistoryOrder_Cancel.java
@@ -1,4 +1,4 @@
-package com.unipet7.mcommerce.fragments.onboarding;
+package com.unipet7.mcommerce.fragments;
 
 import android.os.Bundle;
 
@@ -10,7 +10,7 @@ import android.view.ViewGroup;
 
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.HistoryOrderAdapter;
-import com.unipet7.mcommerce.databinding.FragmentDeliveryOrderBinding;
+import com.unipet7.mcommerce.databinding.FragmentCancelOrderBinding;
 import com.unipet7.mcommerce.models.HistoryOrders;
 
 import java.util.ArrayList;
@@ -18,10 +18,10 @@ import java.util.List;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link HistoryOrder_Delivery#newInstance} factory method to
+ * Use the {@link HistoryOrder_Cancel#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class HistoryOrder_Delivery extends Fragment {
+public class HistoryOrder_Cancel extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -31,11 +31,10 @@ public class HistoryOrder_Delivery extends Fragment {
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
-    FragmentDeliveryOrderBinding binding;
+    FragmentCancelOrderBinding binding;
     HistoryOrderAdapter HistoryAdapter;
     ArrayList<HistoryOrders> Historyorders;
-
-    public HistoryOrder_Delivery() {
+    public HistoryOrder_Cancel() {
         // Required empty public constructor
     }
 
@@ -45,11 +44,11 @@ public class HistoryOrder_Delivery extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment DeliveryOrder.
+     * @return A new instance of fragment CancelOrder.
      */
     // TODO: Rename and change types and number of parameters
-    public static HistoryOrder_Delivery newInstance(String param1, String param2) {
-        HistoryOrder_Delivery fragment = new HistoryOrder_Delivery();
+    public static HistoryOrder_Cancel newInstance(String param1, String param2) {
+        HistoryOrder_Cancel fragment = new HistoryOrder_Cancel();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -69,25 +68,24 @@ public class HistoryOrder_Delivery extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        binding = FragmentDeliveryOrderBinding.inflate(inflater, container, false);
+        binding = FragmentCancelOrderBinding.inflate(inflater,container,false);
         loadData();
         addEvents();
 
-        return binding.getRoot();
-    }
+        return binding.getRoot();    }
 
     private void addEvents() {
     }
 
     private void loadData() {
-        HistoryAdapter = new HistoryOrderAdapter(getActivity(), R.layout.layout_deliverorder, initData());
-        binding.lvlDelOrder.setAdapter(HistoryAdapter);
+        HistoryAdapter = new HistoryOrderAdapter(getActivity(),R.layout.layout_cancelorder, initData());
+        binding.lvlCancelOrder.setAdapter(HistoryAdapter);
     }
 
     private List<HistoryOrders> initData() {
         Historyorders = new ArrayList<>();
-        Historyorders.add(new HistoryOrders("11/12/2024", "1111", "Đang giao", R.drawable.unipet_app_icon, "Royal Canin Rottweiler Puppy con mèo con đi hia long nhong",
-                "Túi", 1.0, 24000.0, 24000.0));
+        Historyorders.add(new HistoryOrders("11/12/2024","1111", "Đã hủy", R.drawable.unipet_app_icon,"Royal Canin Rottweiler Puppy con mèo con đi hia long nhong",
+                "Túi", 1.0,24000.0, 24000.0));
 
         return Historyorders;
     }
