+++ app/src/main/java/com/unipet7/mcommerce/fragments/HistoryOrder_Delivery.java
@@ -1,4 +1,4 @@
-package com.unipet7.mcommerce.fragments.onboarding;
+package com.unipet7.mcommerce.fragments;
 
 import android.os.Bundle;
 
@@ -10,7 +10,7 @@ import android.view.ViewGroup;
 
 import com.unipet7.mcommerce.R;
 import com.unipet7.mcommerce.adapters.HistoryOrderAdapter;
-import com.unipet7.mcommerce.databinding.FragmentConfirmationOrderBinding;
+import com.unipet7.mcommerce.databinding.FragmentDeliveryOrderBinding;
 import com.unipet7.mcommerce.models.HistoryOrders;
 
 import java.util.ArrayList;
@@ -18,10 +18,10 @@ import java.util.List;
 
 /**
  * A simple {@link Fragment} subclass.
- * Use the {@link HistoryOrder_Confirm#newInstance} factory method to
+ * Use the {@link HistoryOrder_Delivery#newInstance} factory method to
  * create an instance of this fragment.
  */
-public class HistoryOrder_Confirm extends Fragment {
+public class HistoryOrder_Delivery extends Fragment {
 
     // TODO: Rename parameter arguments, choose names that match
     // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
@@ -31,10 +31,11 @@ public class HistoryOrder_Confirm extends Fragment {
     // TODO: Rename and change types of parameters
     private String mParam1;
     private String mParam2;
-    FragmentConfirmationOrderBinding binding;
+    FragmentDeliveryOrderBinding binding;
     HistoryOrderAdapter HistoryAdapter;
     ArrayList<HistoryOrders> Historyorders;
-    public HistoryOrder_Confirm() {
+
+    public HistoryOrder_Delivery() {
         // Required empty public constructor
     }
 
@@ -44,11 +45,11 @@ public class HistoryOrder_Confirm extends Fragment {
      *
      * @param param1 Parameter 1.
      * @param param2 Parameter 2.
-     * @return A new instance of fragment ConfirmationOrder.
+     * @return A new instance of fragment DeliveryOrder.
      */
     // TODO: Rename and change types and number of parameters
-    public static HistoryOrder_Confirm newInstance(String param1, String param2) {
-        HistoryOrder_Confirm fragment = new HistoryOrder_Confirm();
+    public static HistoryOrder_Delivery newInstance(String param1, String param2) {
+        HistoryOrder_Delivery fragment = new HistoryOrder_Delivery();
         Bundle args = new Bundle();
         args.putString(ARG_PARAM1, param1);
         args.putString(ARG_PARAM2, param2);
@@ -68,25 +69,25 @@ public class HistoryOrder_Confirm extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        binding = FragmentConfirmationOrderBinding.inflate(inflater,container,false);
+        binding = FragmentDeliveryOrderBinding.inflate(inflater, container, false);
         loadData();
         addEvents();
 
-        return binding.getRoot();    }
+        return binding.getRoot();
+    }
 
     private void addEvents() {
     }
 
     private void loadData() {
-        HistoryAdapter = new HistoryOrderAdapter(getActivity(),R.layout.layout_confirmorder, initData());
-        binding.lvlConfOrder.setAdapter(HistoryAdapter);
+        HistoryAdapter = new HistoryOrderAdapter(getActivity(), R.layout.layout_deliverorder, initData());
+        binding.lvlDelOrder.setAdapter(HistoryAdapter);
     }
 
     private List<HistoryOrders> initData() {
         Historyorders = new ArrayList<>();
-        Historyorders.add(new HistoryOrders("11/12/2024","1111", "Chờ xác nhận", R.drawable.unipet_app_icon,"Royal Canin Rottweiler Puppy con mèo con đi hia long nhong",
-                "Túi", 1.0,24000.0, 24000.0));
+        Historyorders.add(new HistoryOrders("11/12/2024", "1111", "Đang giao", R.drawable.unipet_app_icon, "Royal Canin Rottweiler Puppy con mèo con đi hia long nhong",
+                "Túi", 1.0, 24000.0, 24000.0));
 
         return Historyorders;
     }
