+++ app/src/main/res/drawable/account_ic_bottom_nav.xml
@@ -1,0 +1,15 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="26dp"
+    android:height="25dp"
+    android:viewportWidth="26"
+    android:viewportHeight="25">
+  <group>
+    <clip-path
+        android:pathData="M0.899,0h25v25h-25z"/>
+    <path
+        android:pathData="M10.971,12.599L10.98,12.59L10.988,12.581C11.317,12.191 11.696,11.872 12.128,11.619C12.518,11.39 12.937,11.276 13.399,11.276C13.869,11.276 14.295,11.386 14.687,11.602C15.111,11.836 15.484,12.143 15.811,12.528L15.811,12.529L15.818,12.538C16.274,13.058 16.676,13.618 17.026,14.221C17.425,14.908 17.894,15.548 18.431,16.141L18.431,16.141L18.435,16.145C19.042,16.804 19.653,17.471 20.268,18.147C20.764,18.692 21.002,19.308 21.002,20.026C21.002,20.668 20.8,21.202 20.391,21.665C19.997,22.111 19.538,22.316 18.972,22.316C18.075,22.316 17.187,22.242 16.308,22.093C15.347,21.929 14.377,21.848 13.399,21.848C12.421,21.848 11.451,21.929 10.49,22.093C9.611,22.242 8.723,22.316 7.826,22.316C7.26,22.316 6.802,22.111 6.407,21.665L5.86,22.148L6.407,21.665C5.998,21.202 5.796,20.668 5.796,20.026C5.796,19.308 6.034,18.692 6.53,18.147C7.145,17.472 7.755,16.805 8.361,16.147C8.902,15.569 9.369,14.939 9.761,14.257C10.099,13.67 10.502,13.117 10.971,12.599ZM5.587,11.9C5.046,11.9 4.617,11.721 4.254,11.358C3.891,10.996 3.712,10.566 3.712,10.026C3.712,9.485 3.891,9.056 4.254,8.693C4.617,8.33 5.046,8.152 5.587,8.152C6.127,8.152 6.556,8.33 6.919,8.693C7.282,9.056 7.461,9.485 7.461,10.026C7.461,10.566 7.282,10.996 6.919,11.358C6.556,11.721 6.127,11.9 5.587,11.9ZM10.274,7.733C9.733,7.733 9.304,7.555 8.941,7.192C8.578,6.829 8.4,6.4 8.4,5.859C8.4,5.318 8.578,4.889 8.941,4.526C9.304,4.163 9.733,3.985 10.274,3.985C10.815,3.985 11.244,4.163 11.607,4.526C11.97,4.889 12.148,5.318 12.148,5.859C12.148,6.4 11.97,6.829 11.607,7.192C11.244,7.555 10.815,7.733 10.274,7.733ZM16.524,7.733C15.983,7.733 15.554,7.555 15.191,7.192C14.828,6.829 14.65,6.4 14.65,5.859C14.65,5.318 14.828,4.889 15.191,4.526C15.554,4.163 15.983,3.985 16.524,3.985C17.065,3.985 17.494,4.163 17.857,4.526C18.22,4.889 18.398,5.318 18.398,5.859C18.398,6.4 18.22,6.829 17.857,7.192C17.494,7.555 17.065,7.733 16.524,7.733ZM21.212,11.9C20.671,11.9 20.242,11.721 19.879,11.358C19.516,10.996 19.337,10.566 19.337,10.026C19.337,9.485 19.516,9.056 19.879,8.693C20.242,8.33 20.671,8.152 21.212,8.152C21.752,8.152 22.181,8.33 22.544,8.693C22.907,9.056 23.086,9.485 23.086,10.026C23.086,10.566 22.907,10.996 22.544,11.358C22.181,11.721 21.752,11.9 21.212,11.9Z"
+        android:strokeWidth="1.46"
+        android:fillColor="#00000000"
+        android:strokeColor="#A2A2A2"/>
+  </group>
+</vector>
