+++ app/src/main/res/drawable/love_ic_bottom_nav.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M15.401,3.5C18.753,3.5 21,6.646 21,9.575C21,15.521 11.67,20.389 11.5,20.389C11.33,20.389 2,15.521 2,9.575C2,6.646 4.247,3.5 7.599,3.5C9.514,3.5 10.773,4.455 11.5,5.306C12.227,4.455 13.486,3.5 15.401,3.5Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.46"
+      android:fillColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeColor="#868889"
+      android:strokeLineCap="round"/>
+</vector>
