+++ app/src/main/res/drawable/product_icon_bottom_nav.xml
@@ -1,0 +1,35 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="20dp"
+    android:height="22dp"
+    android:viewportWidth="20"
+    android:viewportHeight="22">
+  <path
+      android:pathData="M14.313,20.5H5.965C2.899,20.5 0.547,19.393 1.215,14.935L1.993,8.894C2.405,6.67 3.823,5.818 5.068,5.818H15.247C16.51,5.818 17.846,6.734 18.322,8.894L19.1,14.935C19.668,18.889 17.379,20.5 14.313,20.5Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.46"
+      android:fillColor="#00000000"
+      android:fillType="evenOdd"
+      android:strokeColor="#A2A2A2"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M14.45,5.599C14.45,3.213 12.516,1.278 10.13,1.278V1.278C8.981,1.273 7.877,1.727 7.063,2.537C6.249,3.348 5.791,4.45 5.791,5.599H5.791"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.46"
+      android:fillColor="#00000000"
+      android:strokeColor="#A2A2A2"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M13.096,10.102H13.05"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.46"
+      android:fillColor="#00000000"
+      android:strokeColor="#A2A2A2"
+      android:strokeLineCap="round"/>
+  <path
+      android:pathData="M7.265,10.102H7.219"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.46"
+      android:fillColor="#00000000"
+      android:strokeColor="#A2A2A2"
+      android:strokeLineCap="round"/>
+</vector>
