+++ app/src/main/res/drawable/shopping_cart_bottom_nav_ic_white.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="30dp"
+    android:height="30dp"
+    android:viewportWidth="30"
+    android:viewportHeight="30">
+  <group>
+    <clip-path
+        android:pathData="M0,0h30v30h-30z"/>
+    <path
+        android:pathData="M8.942,26.947C8.335,26.947 7.819,26.734 7.393,26.309C6.968,25.883 6.755,25.367 6.755,24.76C6.755,24.152 6.968,23.636 7.393,23.21C7.819,22.785 8.335,22.572 8.942,22.572C9.55,22.572 10.066,22.785 10.492,23.21C10.917,23.636 11.13,24.152 11.13,24.76C11.13,25.367 10.917,25.883 10.492,26.309C10.066,26.734 9.55,26.947 8.942,26.947ZM21.058,26.947C20.45,26.947 19.934,26.734 19.508,26.309C19.083,25.883 18.87,25.367 18.87,24.76C18.87,24.152 19.083,23.636 19.508,23.21C19.934,22.785 20.45,22.572 21.058,22.572C21.665,22.572 22.181,22.785 22.607,23.21C23.032,23.636 23.245,24.152 23.245,24.76C23.245,25.367 23.032,25.883 22.607,26.309C22.181,26.734 21.665,26.947 21.058,26.947ZM7.519,7.187L10.688,13.822H19.209C19.281,13.822 19.345,13.804 19.401,13.768C19.458,13.732 19.506,13.682 19.546,13.618L22.899,7.524C22.947,7.436 22.951,7.358 22.911,7.29C22.871,7.221 22.803,7.187 22.707,7.187L7.519,7.187ZM6.62,5.313H23.976C24.487,5.313 24.874,5.53 25.136,5.965C25.398,6.4 25.41,6.845 25.173,7.298L21.168,14.553C20.963,14.913 20.692,15.194 20.354,15.395C20.017,15.597 19.647,15.697 19.245,15.697H10.125L8.678,18.341C8.614,18.438 8.612,18.542 8.672,18.654C8.732,18.766 8.822,18.822 8.942,18.822H22.308C22.574,18.822 22.796,18.912 22.976,19.091C23.155,19.271 23.245,19.493 23.245,19.76C23.245,20.026 23.155,20.248 22.976,20.428C22.796,20.607 22.574,20.697 22.308,20.697H8.942C8.109,20.697 7.483,20.338 7.064,19.619C6.645,18.9 6.63,18.183 7.019,17.466L8.803,14.26L4.255,4.687H2.813C2.546,4.687 2.324,4.598 2.144,4.418C1.965,4.239 1.875,4.016 1.875,3.75C1.875,3.484 1.965,3.261 2.144,3.082C2.324,2.902 2.546,2.813 2.813,2.813H4.724C4.937,2.813 5.135,2.869 5.32,2.982C5.504,3.095 5.644,3.254 5.74,3.459L6.62,5.313Z"
+        android:fillColor="#ffffff"/>
+  </group>
+</vector>
