+++ app/src/main/res/layout/fragment_cancel_order.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.onboarding.HistoryOrder_Cancel"
+    tools:context=".fragments.HistoryOrder_Cancel"
     android:id="@+id/fragmentCancel"
     android:background="@color/white">
