+++ app/src/main/res/layout/fragment_delivery_order.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.onboarding.HistoryOrder_Delivery"
+    tools:context=".fragments.HistoryOrder_Delivery"
     android:id="@+id/fragmentDelivery"
     android:background="@color/white">
