+++ app/src/main/res/layout/fragment_question.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.onboarding.Question">
+    tools:context=".fragments.Question">
 
     <LinearLayout
         android:layout_width="match_parent"
