+++ app/src/main/res/layout/fragment_received_order.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.onboarding.HistoryOrder_Received"
+    tools:context=".fragments.HistoryOrder_Received"
     android:id="@+id/fragmentReceived">
 
     <!-- TODO: Update blank fragment layout -->
