+++ app/src/main/res/values/strings.xml
@@ -33,4 +33,8 @@
     <string name="DiscountVoucher">Voucher Giảm giá</string>
     <string name="ShippingVoucher">Voucher Vận chuyển</string>
     <string name="ButtonApply">XÁC NHẬN</string>
+    <string name="home">Trang Chủ</string>
+    <string name="product_nav">Sản Phẩm</string>
+    <string name="love">Yêu Thích</string>
+    <string name="profile_nav">Tài Khoản</string>
 </resources>
